+++ packages/sui-decorators/src/index.js
@@ -1,7 +1,8 @@
+import {AsyncInlineError} from './decorators/AsyncInlineError/index.js'
 import {cache, invalidateCache} from './decorators/cache/index.js'
 import {Deprecated} from './decorators/deprecated/index.js'
 import inlineError from './decorators/error.js'
 import streamify from './decorators/streamify.js'
 import tracer from './decorators/tracer/index.js'
 
-export {cache, Deprecated, invalidateCache, streamify, inlineError, tracer}
+export {AsyncInlineError, cache, Deprecated, invalidateCache, streamify, inlineError, tracer}