+++ scripts/publish-tagged-packages.mjs
@@ -10,6 +10,7 @@ import path from 'node:path'
 import prettier from 'prettier'
 
 const PACKAGE_REGEX = /packages\/((([a-z]+)-?)+)/ // matches "packages/sui-any-package-name"
+const DEPENDABOT_TAG = 'dependabot'
 
 program
   .name('publish-tagged-packages')
@@ -83,7 +84,8 @@ async function publishPackages() {
   packagesToPublish.forEach(async packageName => {
     const packageConfig = getConfig(packageName)
     const {name} = packageConfig
-    const version = await getPackageVersion({name, tag})
+    const packageTag = tag.startsWith(DEPENDABOT_TAG) ? DEPENDABOT_TAG : tag
+    const version = await getPackageVersion({name, tag: packageTag})
 
     // Set the new tagged version.
     packageConfig.version = version
@@ -99,7 +101,7 @@ async function publishPackages() {
 
     fs.writeFileSync(packageJsonPath, packageJson)
 
-    await $$`npm publish --tag ${tag}`
+    await $$`npm publish --tag ${packageTag}`
   })
 }